Greatly simplified. Removed v1 and v2 version scheme.
This was possible due to the guidance of @dchest and the
realization that the first 32 Bytes of scrypt output will
be identical when dkLen is specified as 32 or 64.

This means we can just *always* generate 64 Bytes of derived
key material and previous username/password combininations
will still generate the same exact keys. All users
now just get the extra signing keys and Base64 keys for free.

Removing code, especially in a security tool, is a good thing.

Good collaboration.